+++ src/disseqt_sdk/validators/composite/__init__.py
@@ -3,4 +3,3 @@
 from .evaluate import CompositeScoreEvaluator
 
 __all__ = ["CompositeScoreEvaluator"]
-