+++ src/disseqt_sdk/validators/output/conceptual_similarity.py
@@ -26,4 +26,3 @@ class ConceptualSimilarityValidator(OutputValidator):
         """Set domain and slug after initialization."""
         object.__setattr__(self, "_domain", ValidatorDomain.OUTPUT_VALIDATION)
         object.__setattr__(self, "_slug", OutputValidation.CONCEPTUAL_SIMILARITY.value)
-