+++ src/disseqt_sdk/validators/output/diversity.py
@@ -25,4 +25,3 @@ class DiversityValidator(OutputValidator):
         """Set domain and slug after initialization."""
         object.__setattr__(self, "_domain", ValidatorDomain.OUTPUT_VALIDATION)
         object.__setattr__(self, "_slug", OutputValidation.DIVERSITY.value)
-