+++ src/disseqt_sdk/validators/output/hate_speech.py
@@ -25,4 +25,3 @@ class OutputHateSpeechValidator(OutputValidator):
         """Set domain and slug after initialization."""
         object.__setattr__(self, "_domain", ValidatorDomain.OUTPUT_VALIDATION)
         object.__setattr__(self, "_slug", OutputValidation.HATE_SPEECH.value)
-