+++ src/disseqt_sdk/validators/output/narrative_continuity.py
@@ -25,4 +25,3 @@ class NarrativeContinuityValidator(OutputValidator):
         """Set domain and slug after initialization."""
         object.__setattr__(self, "_domain", ValidatorDomain.OUTPUT_VALIDATION)
         object.__setattr__(self, "_slug", OutputValidation.NARRATIVE_CONTINUITY.value)
-