+++ src/disseqt_sdk/validators/output/racial_bias.py
@@ -25,4 +25,3 @@ class OutputRacialBiasValidator(OutputValidator):
         """Set domain and slug after initialization."""
         object.__setattr__(self, "_domain", ValidatorDomain.OUTPUT_VALIDATION)
         object.__setattr__(self, "_slug", OutputValidation.RACIAL_BIAS.value)
-