+++ src/disseqt_sdk/validators/output/self_harm.py
@@ -25,4 +25,3 @@ class OutputSelfHarmValidator(OutputValidator):
         """Set domain and slug after initialization."""
         object.__setattr__(self, "_domain", ValidatorDomain.OUTPUT_VALIDATION)
         object.__setattr__(self, "_slug", OutputValidation.SELF_HARM.value)
-